% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembleDownscaleR-package.R
\docType{package}
\name{ensembleDownscaleR-package}
\alias{ensembleDownscaleR}
\alias{ensembleDownscaleR-package}
\title{ensembleDownscaleR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Wyatt Madden \email{wyattgmadden@gmail.com} (\href{https://orcid.org/0000-0002-9792-7077}{ORCID})

}
