% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_fill.R
\name{gap_fill}
\alias{gap_fill}
\title{Make predictions for locations with two observations}
\usage{
gap_fill(grm.pred.1, grm.pred.2, weights)
}
\arguments{
\item{grm.pred.1}{Output from grm_pred() function for first primary variable}

\item{grm.pred.2}{Output from grm_pred() function for second primary variable}

\item{weights}{Output from weight_pred() or ensemble_fit(), depending on if you want ensemble predictions at full prediction locations or observation locations respectively.}
}
\value{
A list including ensemble predictions and standard deviations
}
\description{
This functions uses the weights and predictions from the previous separate grm's to make ensemble predictions
}
\examples{
# gap_fill()


}
