% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grm_pred.R
\name{grm_pred}
\alias{grm_pred}
\title{Make Predictions With Geostatistical Regression Model (GRM)}
\usage{
grm_pred(
  grm.fit,
  X,
  L = NULL,
  M = NULL,
  coords,
  space.id,
  time.id,
  spacetime.id,
  include.additive.spatial.effect = T,
  include.multiplicative.spatial.effect = T,
  n.iter = 500,
  verbose = TRUE,
  in.sample = FALSE
)
}
\arguments{
\item{grm.fit}{Fit object created with grm()}

\item{X}{Standardized primary independent variable vector for all predictions (n_pred)}

\item{L}{Standardized spatial covariate matrix for all predictions (n_pred, p1)}

\item{M}{Standardized spatio-temporal covariate matrix for all predictions (n_pred, p2)}

\item{coords}{Matrix of prediction x, y coodinates (n_pred, 2)}

\item{space.id}{Spatial location ID vector (n)}

\item{time.id}{Temporal location ID vector (n)}

\item{spacetime.id}{ID vector of time points where spatial trends vary (n)}

\item{include.additive.spatial.effect}{Include additive spatial random effects.}

\item{include.multiplicative.spatial.effect}{Include multiplicative spatial random effects.}

\item{n.iter}{Number of iterations used in predictions. Must be <= than post-thined and burned iterations from grm.fit}

\item{verbose}{Print MCMC output}

\item{in.sample}{False if predictions are being made at locations without observations in grm.fit (default is False)}
}
\value{
A data frame containing grm predictions
}
\description{
This function makes predictions using a Bayesian Hierarchical Model (BHM) previously fit with grm()
}
\examples{
# grm_pred()


}
