% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_pred.R
\name{weight_pred}
\alias{weight_pred}
\title{Generate weights from ensemble spatial fit}
\usage{
weight_pred(ensemble.fit, coords, space.id, verbose = TRUE)
}
\arguments{
\item{ensemble.fit}{Fit from ensemble_spatial() function}

\item{coords}{Coordinate matrix for first variable predictions}

\item{space.id}{Space id for first primary variable predictions}

\item{verbose}{Print MCMC output}
}
\value{
A matrix containing weights
}
\description{
This function generates weights using the output from the ensemble_spatial function
}
\examples{
# weight_preds()


}
